@@ -12,22 +12,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^13.0.3",
+    "@angular-devkit/build-angular": "^14.0.4",
     "@angular-material-extensions/password-strength": "^6.0.0",
-    "@angular/animations": "^13.0.2",
+    "@angular/animations": "^14.0.4",
     "@angular/cdk": "^13.0.2",
-    "@angular/cli": "^13.0.3",
-    "@angular/common": "^13.0.2",
-    "@angular/compiler": "^13.0.2",
-    "@angular/compiler-cli": "^13.0.2",
-    "@angular/core": "^13.0.2",
+    "@angular/cli": "^14.0.4",
+    "@angular/common": "^14.0.4",
+    "@angular/compiler": "^14.0.4",
+    "@angular/compiler-cli": "^14.0.4",
+    "@angular/core": "^14.0.4",
     "@angular/flex-layout": "10.0.0-beta.32",
-    "@angular/forms": "^13.0.2",
+    "@angular/forms": "^14.0.4",
     "@angular/http": "^8.0.0-beta.10",
     "@angular/material": "^13.0.2",
-    "@angular/platform-browser": "^13.0.2",
-    "@angular/platform-browser-dynamic": "^13.0.2",
-    "@angular/router": "^13.0.2",
+    "@angular/platform-browser": "^14.0.4",
+    "@angular/platform-browser-dynamic": "^14.0.4",
+    "@angular/router": "^14.0.4",
     "@fortawesome/fontawesome-svg-core": "~1.2.30",
     "@fortawesome/free-brands-svg-icons": "^5.14.0",
     "@fortawesome/free-regular-svg-icons": "^5.14.0",
@@ -61,7 +61,7 @@
   },
   "devDependencies": {
     "@angular-eslint/builder": "^0.8.0-beta.7",
-    "@angular/language-service": "^13.0.2",
+    "@angular/language-service": "^14.0.4",
     "@types/express-serve-static-core": "^4.17.9",
     "@types/file-saver": "^2.0.1",
     "@types/jasmine": "~3.9.1",
