@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-import { AbstractControl, FormControl, Validators } from '@angular/forms'
+import { AbstractControl, UntypedFormControl, Validators } from '@angular/forms'
 import { UserService } from '../Services/user.service'
 import { Component } from '@angular/core'
 import { dom, library } from '@fortawesome/fontawesome-svg-core'
@@ -21,9 +21,9 @@ dom.watch()
   styleUrls: ['./change-password.component.scss']
 })
 export class ChangePasswordComponent {
-  public passwordControl: FormControl = new FormControl('', [Validators.required])
-  public newPasswordControl: FormControl = new FormControl('', [Validators.required, Validators.minLength(5), Validators.maxLength(40)])
-  public repeatNewPasswordControl: FormControl = new FormControl('', [Validators.required, Validators.minLength(5), Validators.maxLength(40), matchValidator(this.newPasswordControl)])
+  public passwordControl: UntypedFormControl = new UntypedFormControl('', [Validators.required])
+  public newPasswordControl: UntypedFormControl = new UntypedFormControl('', [Validators.required, Validators.minLength(5), Validators.maxLength(40)])
+  public repeatNewPasswordControl: UntypedFormControl = new UntypedFormControl('', [Validators.required, Validators.minLength(5), Validators.maxLength(40), matchValidator(this.newPasswordControl)])
   public error: any
   public confirmation: any
 
@@ -75,7 +75,7 @@ export class ChangePasswordComponent {
 }
 
 function matchValidator (newPasswordControl: AbstractControl) {
-  return function matchOtherValidate (repeatNewPasswordControl: FormControl) {
+  return function matchOtherValidate (repeatNewPasswordControl: UntypedFormControl) {
     const password = newPasswordControl.value
     const passwordRepeat = repeatNewPasswordControl.value
     if (password !== passwordRepeat) {
