@@ -5,7 +5,7 @@
 
 import { UserService } from '../Services/user.service'
 import { SecurityQuestionService } from '../Services/security-question.service'
-import { AbstractControl, FormControl, Validators } from '@angular/forms'
+import { AbstractControl, UntypedFormControl, Validators } from '@angular/forms'
 import { Component } from '@angular/core'
 import { dom, library } from '@fortawesome/fontawesome-svg-core'
 import { faSave } from '@fortawesome/free-solid-svg-icons'
@@ -22,10 +22,10 @@ dom.watch()
   styleUrls: ['./forgot-password.component.scss']
 })
 export class ForgotPasswordComponent {
-  public emailControl: FormControl = new FormControl('', [Validators.required, Validators.email])
-  public securityQuestionControl: FormControl = new FormControl({ disabled: true, value: '' }, [Validators.required])
-  public passwordControl: FormControl = new FormControl({ disabled: true, value: '' }, [Validators.required, Validators.minLength(5)])
-  public repeatPasswordControl: FormControl = new FormControl({ disabled: true, value: '' }, [Validators.required, matchValidator(this.passwordControl)])
+  public emailControl: UntypedFormControl = new UntypedFormControl('', [Validators.required, Validators.email])
+  public securityQuestionControl: UntypedFormControl = new UntypedFormControl({ disabled: true, value: '' }, [Validators.required])
+  public passwordControl: UntypedFormControl = new UntypedFormControl({ disabled: true, value: '' }, [Validators.required, Validators.minLength(5)])
+  public repeatPasswordControl: UntypedFormControl = new UntypedFormControl({ disabled: true, value: '' }, [Validators.required, matchValidator(this.passwordControl)])
   public securityQuestion?: string
   public error?: string
   public confirmation?: string
@@ -113,7 +113,7 @@ export class ForgotPasswordComponent {
 }
 
 function matchValidator (passwordControl: AbstractControl) {
-  return function matchOtherValidate (repeatPasswordControl: FormControl) {
+  return function matchOtherValidate (repeatPasswordControl: UntypedFormControl) {
     const password = passwordControl.value
     const passwordRepeat = repeatPasswordControl.value
     if (password !== passwordRepeat) {
