@@ -4,7 +4,7 @@
  */
 
 import { Component, OnInit } from '@angular/core'
-import { FormControl, FormGroup, Validators } from '@angular/forms'
+import { UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms'
 import { mimeType } from './mime-type.validator'
 import { PhotoWallService } from '../Services/photo-wall.service'
 import { IImage } from 'ng-simple-slideshow'
@@ -24,9 +24,9 @@ dom.watch()
 export class PhotoWallComponent implements OnInit {
   public emptyState: boolean = true
   public imagePreview: string
-  public form: FormGroup = new FormGroup({
-    image: new FormControl('', { validators: [Validators.required], asyncValidators: [mimeType] }),
-    caption: new FormControl('', [Validators.required])
+  public form: UntypedFormGroup = new UntypedFormGroup({
+    image: new UntypedFormControl('', { validators: [Validators.required], asyncValidators: [mimeType] }),
+    caption: new UntypedFormControl('', [Validators.required])
   })
 
   public slideshowDataSource: IImage[] = []
