@@ -5,7 +5,7 @@
 
 import { Component, Inject, OnInit } from '@angular/core'
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog'
-import { FormControl, Validators } from '@angular/forms'
+import { UntypedFormControl, Validators } from '@angular/forms'
 import { ProductReviewService } from '../Services/product-review.service'
 import { MatSnackBar } from '@angular/material/snack-bar'
 
@@ -23,7 +23,7 @@ dom.watch()
   styleUrls: ['./product-review-edit.component.scss']
 })
 export class ProductReviewEditComponent implements OnInit {
-  public editReviewControl: FormControl = new FormControl('', [Validators.required, Validators.minLength(1), Validators.maxLength(160)])
+  public editReviewControl: UntypedFormControl = new UntypedFormControl('', [Validators.required, Validators.minLength(1), Validators.maxLength(160)])
   public error: string | null = null
 
   constructor (@Inject(MAT_DIALOG_DATA) public data: { reviewData: Review }, private readonly productReviewService: ProductReviewService, private readonly dialogRef: MatDialogRef<ProductReviewEditComponent>,
