@@ -4,7 +4,7 @@
  */
 
 import { Component, NgZone } from '@angular/core'
-import { FormControl, FormGroup, Validators } from '@angular/forms'
+import { UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms'
 import { TwoFactorAuthService } from '../Services/two-factor-auth-service'
 import { CookieService } from 'ngx-cookie'
 import { UserService } from '../Services/user.service'
@@ -25,8 +25,8 @@ interface TokenEnterFormFields {
   styleUrls: ['./two-factor-auth-enter.component.scss']
 })
 export class TwoFactorAuthEnterComponent {
-  public twoFactorForm: FormGroup = new FormGroup({
-    token: new FormControl('', [Validators.minLength(6), Validators.maxLength(6), Validators.required, Validators.pattern('^[\\d]{6}$')])
+  public twoFactorForm: UntypedFormGroup = new UntypedFormGroup({
+    token: new UntypedFormControl('', [Validators.minLength(6), Validators.maxLength(6), Validators.required, Validators.pattern('^[\\d]{6}$')])
   })
 
   public errored: Boolean = false
