@@ -5,7 +5,7 @@
 
 import { Component, NgZone, OnInit } from '@angular/core'
 import { WalletService } from '../Services/wallet.service'
-import { FormControl, Validators } from '@angular/forms'
+import { UntypedFormControl, Validators } from '@angular/forms'
 import { Router } from '@angular/router'
 
 @Component({
@@ -15,7 +15,7 @@ import { Router } from '@angular/router'
 })
 export class WalletComponent implements OnInit {
   public balance: string
-  public balanceControl: FormControl = new FormControl('', [Validators.required, Validators.min(10), Validators.max(1000)])
+  public balanceControl: UntypedFormControl = new UntypedFormControl('', [Validators.required, Validators.min(10), Validators.max(1000)])
 
   constructor (private readonly router: Router, private readonly walletService: WalletService, private readonly ngZone: NgZone) { }
 
